+++ projects/core/src/page/title/facade/index.ts
@@ -1,7 +1,2 @@
 export * from './page-title.service';
-export * from './page-title.resolver';
-export * from './resolvers/product-page-title.resolver';
-export * from './resolvers/category-page-title.resolver';
-export * from './resolvers/content-page-title.resolver';
-export * from './resolvers/search-page-title.resolver';
-export * from './resolvers/checkout-page-title.resolver';
+export * from '../resolvers/page-title.resolver';
